+++ code/Javascript/promise-without-promise/my-promise.js
@@ -1,0 +1,18 @@
+const STATUS = {
+  pending: 0,
+  resolved: 1,
+  rejected: 2,
+}
+
+class MyPromise {
+
+  constructor () {
+
+  }
+
+}
+
+
+module.exports = {
+  MyPromise
+}