+++ update.sh
@@ -10,5 +10,6 @@ then
 fi
 
 git add . && git commit -m 'update'
+git status
 
 git push origin $CURRENT_BRANCH